+++ Main.storyboard
@@ -120,10 +120,10 @@
             </objects>
             <point key="canvasLocation" x="2012" y="-350"/>
         </scene>
-        <!--Main Tab View Controller-->
+        <!--Main Tab Bar Controller-->
         <scene sceneID="Vab-cG-Dvg">
             <objects>
-                <tabBarController automaticallyAdjustsScrollViewInsets="NO" id="IzG-bR-w03" customClass="MainTabViewController" customModule="ManageClothesUIkit" customModuleProvider="target" sceneMemberID="viewController">
+                <tabBarController automaticallyAdjustsScrollViewInsets="NO" id="IzG-bR-w03" customClass="MainTabBarController" customModule="ManageClothesUIkit" customModuleProvider="target" sceneMemberID="viewController">
                     <toolbarItems/>
                     <tabBar key="tabBar" contentMode="scaleToFill" insetsLayoutMarginsFromSafeArea="NO" id="HJN-jv-QAN">
                         <rect key="frame" x="0.0" y="0.0" width="1000" height="1000"/>
